move footer tests to their own file
